Move writer to its own module
