--- tests/dict/en/ru/temp-ipa/dict/term_meta_bank_1.json
@@ -1,17 +1,0 @@
-[
-  [
-    "read",
-    "ipa",
-    {
-      "reading": "read",
-      "transcriptions": [
-        {
-          "ipa": "[rɪːd]"
-        },
-        {
-          "ipa": "[rɛd]"
-        }
-      ]
-    }
-  ]
-]
